+++ rustfmt.toml
@@ -1,0 +1,1 @@
+max_width = 120
